python_source/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.3
xlwt==1.3.0
openpyxl==3.1.2
